fix(dockerclient): invalid memory address
In general, the convention in Go is to assume that if there's an error, there's no response, therefore no body to be closed. Unfortunately, this can't be guaranteed by the type system, but we are using a package from the standard library to send requests, so we should be safe.